expose the tabs from the new data/ folder
